+++ MKE/Graphics/MKE/Texture.cpp
@@ -39,7 +39,7 @@ void mk::Texture::bind(const Texture* texture) {
 		glBindTexture(GL_TEXTURE_2D, 0);
 }
 
-void mk::Texture::loadFromMemory(usize width, usize height, const void* data, GLenum format) {
+void mk::Texture::loadFromMemory(u32 width, u32 height, const void* data, GLenum format) {
 	size = { width, height };
 	glBindTexture(GL_TEXTURE_2D, texture_id);
